requirements.txt
web3==5.5.0
geohash2==1.1.0
requests==2.23.0
Pillow==7.1.1
python-dotenv==0.12.0